+++ src/protocol/chromium-bidi.ts
@@ -15,8 +15,8 @@
  * limitations under the License.
  */
 
-import type * as WebDriverBidi from './cddl/webdriver-bidi.js';
 import type * as Cdp from './cdp.js';
+import type * as WebDriverBidi from './generated/webdriver-bidi.js';
 
 export type EventNames =
   // keep-sorted start
