+++ src/protocol/protocol.ts
@@ -16,5 +16,5 @@
  */
 export * as Cdp from './cdp.js';
 export * as ChromiumBidi from './chromium-bidi.js';
-export * from './cddl/webdriver-bidi.js';
+export * from './generated/webdriver-bidi.js';
 export * from './ErrorResponse.js';
